Backend/requirements.txt
requests
sentence-transformers>=2.3.1
fastapi
uvicorn
pydantic
rich
tinydb
python-dotenv
